+++ data/beaconrestapi/src/integration-test/java/tech/pegasys/teku/beaconrestapi/v1/node/GetIdentityIntegrationTest.java
@@ -22,6 +22,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 import okhttp3.Response;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import tech.pegasys.teku.api.response.v1.node.Identity;
 import tech.pegasys.teku.api.response.v1.node.IdentityResponse;
@@ -30,30 +31,60 @@ import tech.pegasys.teku.beaconrestapi.AbstractDataBackedRestAPIIntegrationTest;
 import tech.pegasys.teku.beaconrestapi.handlers.v1.node.GetIdentity;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.networking.p2p.mock.MockNodeId;
+import tech.pegasys.teku.spec.SpecMilestone;
 import tech.pegasys.teku.spec.datastructures.networking.libp2p.rpc.metadata.MetadataMessage;
 import tech.pegasys.teku.spec.util.DataStructureUtil;
 
 public class GetIdentityIntegrationTest extends AbstractDataBackedRestAPIIntegrationTest {
 
   private final DataStructureUtil dataStructureUtil = new DataStructureUtil(spec);
+  private final String enr = "enr";
+  private final String address = "address";
+  private final String discoveryAddress = "discoveryaddress";
+  private final MockNodeId node1 = new MockNodeId(0);
+  private final UInt64 seqnr = dataStructureUtil.randomUInt64();
+
+  @BeforeEach
+  void setup() {
+    when(eth2P2PNetwork.getNodeId()).thenReturn(node1);
+    when(eth2P2PNetwork.getEnr()).thenReturn(Optional.of(enr));
+    when(eth2P2PNetwork.getNodeAddress()).thenReturn(address);
+    when(eth2P2PNetwork.getDiscoveryAddress()).thenReturn(Optional.of(discoveryAddress));
+  }
 
   @Test
   public void shouldReturnNetworkIdentity() throws Exception {
     startRestAPIAtGenesis();
-    String enr = "enr";
-    String address = "address";
-    String discoveryAddress = "discoveryaddress";
-    final MockNodeId node1 = new MockNodeId(0);
-    final UInt64 seqnr = dataStructureUtil.randomUInt64();
     final MetadataMessage metadataMessage =
         spec.getGenesisSchemaDefinitions()
             .getMetadataMessageSchema()
             .create(seqnr, List.of(1, 11, 15), Collections.emptyList());
 
-    when(eth2P2PNetwork.getNodeId()).thenReturn(node1);
-    when(eth2P2PNetwork.getEnr()).thenReturn(Optional.of(enr));
-    when(eth2P2PNetwork.getNodeAddress()).thenReturn(address);
-    when(eth2P2PNetwork.getDiscoveryAddress()).thenReturn(Optional.of(discoveryAddress));
+    when(eth2P2PNetwork.getMetadata()).thenReturn(metadataMessage);
+
+    final Response response = get();
+    assertThat(response.code()).isEqualTo(SC_OK);
+    final IdentityResponse identityResponse =
+        jsonProvider.jsonToObject(response.body().string(), IdentityResponse.class);
+    assertThat(identityResponse.data)
+        .isEqualTo(
+            new Identity(
+                node1.toBase58(),
+                enr,
+                List.of(address),
+                List.of(discoveryAddress),
+                new Metadata(metadataMessage)));
+  }
+
+  @Test
+  public void shouldReturnNetworkIdentityAltair() throws Exception {
+    startRestAPIAtGenesis(SpecMilestone.ALTAIR);
+
+    final MetadataMessage metadataMessage =
+        spec.getGenesisSchemaDefinitions()
+            .getMetadataMessageSchema()
+            .create(seqnr, List.of(1, 11, 15), List.of(0, 1, 2, 3));
+
     when(eth2P2PNetwork.getMetadata()).thenReturn(metadataMessage);
 
     final Response response = get();